+++ client/src/main/java/com/streamr/client/subs/BasicSubscription.java
@@ -7,7 +7,6 @@ import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
-import com.streamr.client.utils.DecryptionQueues;
 import com.streamr.client.stream.GroupKey;
 import com.streamr.client.stream.GroupKeyStore;
 import com.streamr.client.stream.KeyExchangeUtil;
